move getContentPageReferences to utils
