+++ src/app/content/hooks/locationChange/resolveContent.ts
@@ -10,11 +10,11 @@ import * as select from '../../selectors';
 import { ArchivePage, Book, PageReferenceError, PageReferenceMap } from '../../types';
 import {
   formatBookData,
+  getContentPageReferences,
   getIdFromPageParam,
   getPageIdFromUrlParam,
 } from '../../utils';
 import { archiveTreeContainsNode, archiveTreeSectionIsBook } from '../../utils/archiveTreeUtils';
-import { stripIdVersion } from '../../utils/idUtils';
 import { getUrlParamForPageId, getUrlParamsForBook } from '../../utils/urlUtils';
 
 export default async(
@@ -30,29 +30,6 @@ export default async(
 
   return {book, page};
 };
-
-export function getContentPageReferences(htmlContent: string) {
-  const matches = (htmlContent.match(/.\/([a-z0-9-]+(@[\d.]+)?):([a-z0-9-]+.xhtml)/g) || [])
-    .map((match) => {
-      const [bookMatch, pageMatch] = match.split(':');
-      const pageId = pageMatch.split('.xhtml')[0];
-      let [bookId, bookVersion] = bookMatch.split('@') as [string, string | undefined];
-      bookId = bookId.substr(2);
-
-      if (!bookVersion && BOOKS[bookId]) {
-        bookVersion = BOOKS[bookId].defaultVersion;
-      }
-
-      return {
-        bookId,
-        bookVersion,
-        match,
-        pageId: stripIdVersion(pageId),
-      };
-    });
-
-  return matches;
-}
 
 const getBookResponse = async(
   osWebLoader: AppServices['osWebLoader'],
@@ -167,20 +144,28 @@ const resolvePage = async(
   }
 };
 
+const getInputReferenceInfo = (bookId: string, inputVersion?: string) => {
+  const defaultVersion = BOOKS[bookId] ? BOOKS[bookId].defaultVersion : undefined;
+  const bookVersion = inputVersion ? inputVersion : defaultVersion;
+  return {bookId, bookVersion};
+};
+
 export const getBookInformation = async(
   services: AppServices & MiddlewareAPI,
   reference: ReturnType<typeof getContentPageReferences>[number]
 ) => {
-  const osWebBook =  await services.osWebLoader.getBookFromId(reference.bookId);
-  const archiveBook = await services.archiveLoader.book(
-    reference.bookId, reference.bookVersion || 'undefined'
-  ).load().catch((error) => {
-    if (UNLIMITED_CONTENT) {
-      return undefined;
-    } else {
-      throw error;
-    }
-  });
+  const { bookId, bookVersion } = getInputReferenceInfo(reference.bookId, reference.bookVersion);
+
+  const osWebBook =  await services.osWebLoader.getBookFromId(bookId);
+  const archiveBook = await services.archiveLoader.book(bookId, bookVersion || 'undefined')
+    .load()
+    .catch((error) => {
+      if (UNLIMITED_CONTENT) {
+        return undefined;
+      } else {
+        throw error;
+      }
+    });
 
   if (archiveBook && archiveTreeSectionIsBook(archiveBook.tree)) {
     return {osWebBook, archiveBook};
